Combine into one page, and add Metadata support.
